server/requirements.txt
anthropic==0.68.0
python-dotenv==1.1.1
requests==2.32.5
psycopg2-binary==2.9.9
schedule==1.2.0
fastapi==0.104.1
uvicorn==0.24.0